Commit for Leo Treggiari:
44568 - [Managed Build] -Xlinker option requires space separator
80119 - [Managed Build] Error in the Xlinker option's generated output
The code and the manifest file have been changed to correctly deal with
the -Xlinker option.  Multiple entries have separate -Xlinker options,
and there is a space between -Xlinker and the value.  The space is
handled by the new option.command functionality - "${VALUE}".

77399 - Managed Make Builder mangles subdir.mk if configuration of
        linked resource was changed
This was partially fixed before and was partially a user error.
Code has been added to output an error message to the console when
MBS sees a duplicate identifier in the loaded manifest files.

Partial fix:
80067 - [Managed Build] Wrong command for building in MMS
A fix has been added so that a command is not stored with a Tool
unless the user changes the value - i.e the Tool will inherit the
value from its suoer-class.  There is still an error with the Gnu
makefile generator when a configuration tool and a resource
configuration tool have different commands specified by the user.
This will be fixed later.